Widen flat-ride rotation animation to uint16_t and add Freestyle ride type
Widens flatRideAnimationFrame/TimeToSpriteMap/FramesPerDir from uint8_t to
uint16_t (0xFFFF sentinel) so FramesPerDir can exceed 255, and adds
FlatRideAnimationPhase::ResetRotationsOnEntry so a program can have multiple
independent RepeatUntilRotationsComplete phases. Registers the new Freestyle
ride type (RIDE_TYPE_FREESTYLE, 6x6 footprint, 1024 frames/dir, 24 riders)
with its 9-phase "Long Program" (kFreestylePhases/kFreestylePrograms).

+++ src/openrct2/paint/track/thrill/GenericFlatRide.cpp
@@ -48,7 +48,7 @@ static void PaintGenericRotatingStructure(
     if (baseImageId == 0)
         return;
 
-    uint8_t animFrame = 0;
+    uint16_t animFrame = 0;
     Vehicle* vehicle = nullptr;
     if (ride.flags.has(RideFlag::onTrack) && !ride.vehicles[0].IsNull())
     {
